Tried to update jekyll and ran bundle update - same
It infinitely nests where? On the generation of the feed or somewhere
else?

The error log mentions index#excerpt - whatever that is. I tried to
comment out excerpts - still same.

I tried to roll back to older commits -- same. Is it a Ruby thing then?

Also, I am trying to run bundle exec jekyll serve in a fresh
installation.

	modified:   Gemfile
	modified:   Gemfile.lock